+++ pubspec.yaml
@@ -29,6 +29,7 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.3
   keyboard_visibility: any
+  provider: any
 
 dev_dependencies:
   flutter_test:
